Updates card component
