+++ main/core/Resources/modules/administration/logs/connection/components/connection-list.jsx
@@ -2,6 +2,7 @@ import {LINK_BUTTON} from '#/main/app/buttons'
 
 import {trans} from '#/main/core/translation'
 
+import {convertTimestampToString} from '#/main/core/administration/logs/connection/utils'
 import {LogConnectPlatformCard} from '#/main/core/administration/logs/connection/data/components/log-connect-platform-card'
 
 const ConnectionList = {
@@ -34,23 +35,7 @@ const ConnectionList = {
       label: trans('duration'),
       displayed: true,
       filterable: false,
-      calculated: (rowData) => {
-        let result = null
-        let duration = rowData.duration
-
-        if (duration !== null) {
-          const hours = Math.floor(duration / 3600)
-          duration %= 3600
-          const minutes = Math.floor(duration / 60)
-          const seconds = duration % 60
-
-          result = `${hours}:`
-          result += 10 > minutes ? `0${minutes}:` : `${minutes}:`
-          result += 10 > seconds ? `0${seconds}` : `${seconds}`
-        }
-
-        return result
-      },
+      calculated: (rowData) => rowData.duration !== null ? convertTimestampToString(rowData.duration) : null
     }
   ],
   card: LogConnectPlatformCard
