+++ main/core/Resources/modules/administration/logs/connection/data/components/log-connect-platform-card.jsx
@@ -3,8 +3,9 @@ import {PropTypes as T} from 'prop-types'
 
 import {asset} from '#/main/core/scaffolding/asset'
 import {DataCard} from '#/main/core/data/components/data-card'
-
-// import {LogConnectPlatform as LogConnectPlatformType} from '#/main/core/administration/logs/connection/prop-types'
+import {UserAvatar} from '#/main/core/user/components/avatar'
+import {convertTimestampToString} from '#/main/core/administration/logs/connection/utils'
+import {LogConnectPlatform as LogConnectPlatformType} from '#/main/core/administration/logs/connection/prop-types'
 
 const LogConnectPlatformCard = props =>
   <DataCard
@@ -12,18 +13,13 @@ const LogConnectPlatformCard = props =>
     id={props.data.id}
     poster={props.data.user.thumbnail ? asset(props.data.user.thumbnail) : null}
     icon={<UserAvatar picture={props.data.user.picture} alt={true} />}
-    title={props.data.user.username}
-    subtitle={props.data.user.firstName + ' ' + props.data.user.lastName}
-    contentText={props.data.user.meta.description}
-    footer={
-      <span>
-        {props.data.duration}
-      </span>
-    }
+    title={props.data.user.firstName + ' ' + props.data.user.lastName}
+    subtitle={props.data.date}
+    contentText={props.data.duration !== null ? convertTimestampToString(props.data.duration) : null}
   />
 
 LogConnectPlatformCard.propTypes = {
-  // data: T.shape(LogConnectPlatformType.propTypes).isRequired
+  data: T.shape(LogConnectPlatformType.propTypes).isRequired
 }
 
 export {
